[SDWebImage] Update to version 3.5.4

+++ SDWebImage/samples/SDWebImageSimpleSample/SDWebImageSimpleSample/SampleViewController.cs
@@ -32,7 +32,7 @@ namespace SDWebImageSimpleSample
 			);	
 		}
 
-		void ProgressHandler (uint receivedSize, long expectedSize)
+		void ProgressHandler (int receivedSize, int expectedSize)
 		{
 			if (expectedSize > 0) {
 				InvokeOnMainThread (()=> {
